enigma2 (20120321 rel32 -> 20120327 rel32) ------------------------------------------ - fixed problem on short signal loss when recording with linked tuners   (which made switching channels impossible until the recording was finished) - fixed skipping forward/backward on some MPEG2-content (> 16X) - fixed possible crash in device manager - added possibility to remove all terrestrial / cable services in channel selection context menu - updated default favourite lists - added 7020hd to common interface assignment plugin prerequesites

+++ usr/include/enigma2/lib/dvb/pvrparse.h
@@ -50,7 +50,7 @@ public:
 	
 	bool empty();
 	
-	typedef unsigned long long structure_data;
+	typedef off_t structure_data;
 		/* this is usually:
 			sc | (other_information << 8)
 			but is really specific to the used video encoder.
@@ -60,11 +60,11 @@ public:
 		/* get a structure entry at given offset (or previous one, if no exact match was found).
 		   optionall, return next element. Offset will be returned. this allows you to easily 
 		   get previous and next structure elements. */
-	int getStructureEntry(off_t &offset, unsigned long long &data, int get_next);
+	int getStructureEntry(off_t &offset, off_t &data, int get_next);
 
 	std::string m_filename;
 	int m_structure_cache_valid;
-	unsigned long long m_structure_cache[1024];
+	off_t m_structure_cache[1024];
 	FILE *m_structure_read, *m_structure_write;
 };
